WordListActivity

+++ app/src/main/java/com/donutsbite/godofmem/feature/book/data/Book.kt
@@ -3,8 +3,8 @@ package com.donutsbite.godofmem.feature.book.data
 import com.donutsbite.godofmem.api.response.BookResponse
 
 data class Book(
-    val id: Int,
-    val userId: Int,
+    val id: Long,
+    val userId: Long,
     val title: String,
     val description: String?
 ) {
